+++ examples/src/main/java/pfg/kraken_examples/ExampleEurobot2017.java
@@ -108,6 +108,8 @@ public class ExampleEurobot2017
 				System.out.println(p);
 			}
 			
+			System.out.println(kraken.getTentaclesStatistics());
+
 			display.refresh();
 		}
 		catch(PathfindingException e)